tests/patterntest.c (test_compilation) A couple of 64-bit printf format
Tue Jun 11 17:03:39 2002  Owen Taylor  <otaylor@redhat.com>

        * tests/patterntest.c (test_compilation)
        * glib/gmem.c (profiler_try_realloc): A couple
        of 64-bit printf format fix from George Lebl.

        * glib/gconvert.c (open_converter): Fix gsize/gint
        mixup for g_iconv() arguments.

+++ glib/gconvert.c
@@ -358,9 +358,9 @@ open_converter (const gchar *to_codeset,
 	   * NULL for anything but inbuf; work around that. (NULL outbuf
 	   * or NULL *outbuf is allowed by Unix98.)
 	   */
-	  gint inbytes_left = 0;
+	  gsize inbytes_left = 0;
 	  gchar *outbuf = NULL;
-	  gint outbytes_left = 0;
+	  gsize outbytes_left = 0;
 		
           cd = bucket->cd;
           bucket->used = TRUE;
